feat(journey-maps): add several new examples for journey maps client. delete full example
