+++ src/components-examples/journey-maps/angular/journey-maps-basic/journey-maps-basic-example.ts
@@ -9,7 +9,7 @@ declare global {
 }
 
 /**
- * @title Journey Maps Basic Example
+ * @title Journey Maps - Basic Example
  */
 @Component({
   selector: 'sbb-journey-maps-basic-example',
