--- src/components-examples/journey-maps/angular/journey-maps-full/journey-maps-full-example.css
@@ -1,39 +1,0 @@
-@import 'https://unpkg.com/maplibre-gl@3.6.1/dist/maplibre-gl.css';
-
-sbb-notification.api-key-info {
-  margin-bottom: 30px;
-}
-
-ul {
-  margin: 9px 0;
-  padding: 0;
-}
-
-li {
-  margin: 0;
-  padding: 5px 0;
-  list-style-type: none;
-}
-
-li button {
-  margin: 3px 9px 3px 0;
-}
-
-sbb-form-field {
-  padding-right: 1em;
-}
-
-.web-map {
-  width: 100%;
-  height: 500px;
-  min-height: 500px;
-  box-shadow: 0 5px 5px -5px rgba(0, 0, 0, 0.15);
-}
-
-.rail-options {
-  margin-top: 5px;
-}
-
-.rail-options .sbb-select {
-  width: 305px;
-}
