+++ src/components-examples/journey-maps/angular/journey-maps-map-navigation/journey-maps-map-navigation-example.css
@@ -1,0 +1,90 @@
+@import 'https://unpkg.com/maplibre-gl@3.6.1/dist/maplibre-gl.css';
+
+sbb-notification.api-key-info {
+  margin-bottom: 30px;
+}
+
+.web-map {
+  width: 100%;
+  height: 500px;
+  min-height: 500px;
+  box-shadow: 0 5px 5px -5px rgba(0, 0, 0, 0.15);
+}
+
+.sbb-navigation {
+  display: flex;
+  flex-direction: column;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  gap: 5rem;
+}
+
+.sbb-navigation .row-1 {
+  margin-top: 1rem;
+  display: flex;
+  justify-content: space-around;
+}
+
+.sbb-navigation .row-2 {
+  display: flex;
+  justify-content: start;
+  gap: 4rem;
+  margin-bottom: 2rem;
+}
+
+.sbb-navigation li {
+  list-style: none;
+}
+
+.sbb-navigation button {
+  margin: 0 5px;
+}
+
+.sbb-navigation .sbb-label {
+  white-space: nowrap;
+  padding: 0 10px;
+}
+
+.map-controls {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  margin-top: 5px;
+}
+
+.map-controls button[sbb-ghost-button] {
+  position: absolute;
+  margin: 0;
+}
+
+/* North button */
+.map-controls button[sbb-ghost-button]:nth-child(1) {
+  top: 0;
+  transform: translateY(-100%);
+}
+
+/* East button */
+.map-controls button[sbb-ghost-button]:nth-child(3) {
+  right: 0;
+  transform: translateX(100%);
+}
+
+/* South button */
+.map-controls button[sbb-ghost-button]:nth-child(4) {
+  bottom: 0;
+  transform: translateY(100%);
+}
+
+/* West button */
+.map-controls button[sbb-ghost-button]:nth-child(5) {
+  left: 0;
+  transform: translateX(-100%);
+}
+
+.icon-container {
+  z-index: 1;
+  margin: 1rem;
+}
