+++ src/components-examples/journey-maps/angular/journey-maps-pois-markers/journey-maps-pois-markers-example.css
@@ -1,0 +1,63 @@
+@import 'https://unpkg.com/maplibre-gl@3.6.1/dist/maplibre-gl.css';
+
+sbb-notification.api-key-info {
+  margin-bottom: 30px;
+}
+
+.web-map {
+  width: 100%;
+  height: 500px;
+  min-height: 500px;
+  box-shadow: 0 5px 5px -5px rgba(0, 0, 0, 0.15);
+}
+
+.poi-controls {
+  display: flex;
+  flex-direction: column;
+  list-style: none;
+  margin-top: 1rem;
+  gap: 5rem;
+}
+
+.sbb-form-field:first-child {
+  width: 100%;
+  max-width: 400px;
+}
+
+.poi-controls form > div {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+div[formGroupName='pointsOfInterestOptions'] {
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  gap: 1rem;
+}
+
+li[formGroupName='selectedPoi'] .class-of-first-element {
+  margin-right: 1rem;
+}
+
+.search-poi {
+  display: flex;
+  align-items: start;
+  gap: 0.5rem;
+}
+
+.row-1,
+.row-2 {
+  display: flex;
+  justify-content: start;
+  gap: 3.5rem;
+  flex-wrap: nowrap;
+}
+.select-poi-button {
+  margin-left: 1rem;
+}
+.unselect-all-button {
+  margin-left: 1rem;
+  margin-right: 1rem;
+}
