+++ src/components-examples/journey-maps/angular/journey-maps-routes-zones/journey-maps-routes-zones-example.css
@@ -1,0 +1,26 @@
+@import 'https://unpkg.com/maplibre-gl@3.6.1/dist/maplibre-gl.css';
+
+sbb-notification.api-key-info {
+  margin-bottom: 30px;
+}
+
+.web-map {
+  width: 100%;
+  height: 500px;
+  min-height: 500px;
+  box-shadow: 0 5px 5px -5px rgba(0, 0, 0, 0.15);
+}
+
+.poi-controls {
+  display: flex;
+  flex-direction: column;
+  list-style: none;
+  margin-top: 1rem;
+  gap: 5rem;
+}
+
+.poi-controls form > div {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
