+++ src/components-examples/journey-maps/angular/journey-maps-ui-options/journey-maps-ui-options-example.css
@@ -1,0 +1,23 @@
+@import 'https://unpkg.com/maplibre-gl@3.6.1/dist/maplibre-gl.css';
+
+sbb-notification.api-key-info {
+  margin-bottom: 30px;
+}
+
+.web-map {
+  width: 100%;
+  height: 500px;
+  min-height: 500px;
+  box-shadow: 0 5px 5px -5px rgba(0, 0, 0, 0.15);
+}
+
+form {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  padding: 20px 0;
+}
+
+sbb-checkbox {
+  margin: 8px 0;
+}
