+++ src/components-examples/journey-maps/angular/shared/journey/be-sh.ts
