+++ src/components-examples/journey-maps/angular/shared/journey/zh-be_wyleregg.ts
