+++ src/components-examples/journey-maps/angular/shared/journey/zh-sh_waldfriedhof.ts
