+++ src/components-examples/journey-maps/angular/shared/markers.ts
