+++ src/components-examples/journey-maps/angular/shared/routes/biel-lyss.ts
