+++ src/components-examples/journey-maps/angular/shared/routes/bn-ls.ts
