+++ src/components-examples/journey-maps/angular/shared/transfer/bern-indoor.ts
