+++ src/components-examples/journey-maps/angular/shared/transfer/geneve-indoor.ts
