+++ src/components-examples/journey-maps/angular/shared/transfer/zurich-indoor.ts
