+++ src/components-examples/journey-maps/angular/shared/zone/bern-burgdorf.ts
